Force Format Change
The current Force Format functionality in the code simply changes the file extension. This is not good when a user uploads a large PNG and you require to save a much more compressed JPG. This edit flattens RGBA images into RGB if force_format is set to JPEG, JPG, jpeg or jpg.